move m3u files into dedicated directory.
These playlist files are useful only to the installation at
https://asteroid.radio, because they reference tracks that likely do
not exist in any other library where this code might be run.
Regardless, it's useful to me, and saving this work somewhere durable
has intrinsic value to me, operationally. If nothing else, this shows
the structure of a plalist for use with this code.